requirements.txt
Django==3.2
django-allauth==0.50.0
djangorestframework==3.13.1
oauthlib==3.2.0
Pillow
psycopg2-binary
python-decouple==3.6
requests
requests-oauthlib
six
stripe==3.2.0
uWSGI==2.0.17
django-storages==1.12.3
boto3
botocore
django-crontab==0.7.1
PyJWT
cryptography==2.8
twilio
bleach